Added the console node
Added the console node. This will allow the user to be on a different
computer than the one's running the database and still be able to
execute commands. This also opens the possibility to setup a
user/password system so that only certain logins can actually connect to
a database.